Move blockstore visitors to remove lmdb dependency (#1186)
