+++ rai/secure/blockstore.hpp
@@ -4,6 +4,70 @@
 
 namespace rai
 {
+class transaction;
+class block_store;
+/**
+ * Determine the balance as of this block
+ */
+class balance_visitor : public rai::block_visitor
+{
+public:
+	balance_visitor (rai::transaction const &, rai::block_store &);
+	virtual ~balance_visitor () = default;
+	void compute (rai::block_hash const &);
+	void send_block (rai::send_block const &) override;
+	void receive_block (rai::receive_block const &) override;
+	void open_block (rai::open_block const &) override;
+	void change_block (rai::change_block const &) override;
+	void state_block (rai::state_block const &) override;
+	rai::transaction const & transaction;
+	rai::block_store & store;
+	rai::block_hash current_balance;
+	rai::block_hash current_amount;
+	rai::uint128_t balance;
+};
+
+/**
+ * Determine the amount delta resultant from this block
+ */
+class amount_visitor : public rai::block_visitor
+{
+public:
+	amount_visitor (rai::transaction const &, rai::block_store &);
+	virtual ~amount_visitor () = default;
+	void compute (rai::block_hash const &);
+	void send_block (rai::send_block const &) override;
+	void receive_block (rai::receive_block const &) override;
+	void open_block (rai::open_block const &) override;
+	void change_block (rai::change_block const &) override;
+	void state_block (rai::state_block const &) override;
+	void from_send (rai::block_hash const &);
+	rai::transaction const & transaction;
+	rai::block_store & store;
+	rai::block_hash current_amount;
+	rai::block_hash current_balance;
+	rai::uint128_t amount;
+};
+
+/**
+ * Determine the representative for this block
+ */
+class representative_visitor : public rai::block_visitor
+{
+public:
+	representative_visitor (rai::transaction const & transaction_a, rai::block_store & store_a);
+	virtual ~representative_visitor () = default;
+	void compute (rai::block_hash const & hash_a);
+	void send_block (rai::send_block const & block_a) override;
+	void receive_block (rai::receive_block const & block_a) override;
+	void open_block (rai::open_block const & block_a) override;
+	void change_block (rai::change_block const & block_a) override;
+	void state_block (rai::state_block const & block_a) override;
+	rai::transaction const & transaction;
+	rai::block_store & store;
+	rai::block_hash current;
+	rai::block_hash result;
+};
 template <typename T, typename U>
 class store_iterator_impl
 {
